src/main/java/com/williammacedo/bookstoremanager/author/controller/AuthorController.java
package com.williammacedo.bookstoremanager.author.controller;

import com.williammacedo.bookstoremanager.author.dto.AuthorDTO;
import com.williammacedo.bookstoremanager.author.service.AuthorService;
import lombok.AllArgsConstructor;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import javax.validation.Valid;
import java.util.List;

@AllArgsConstructor
@RestController
@RequestMapping("v1/authors")
public class AuthorController implements AuthorControllerDocs {

    private AuthorService service;

    @GetMapping("/{id}")
    public AuthorDTO findById(@PathVariable Long id) {
        return service.findById(id);
    }

    @GetMapping
    public List<AuthorDTO> findAll() {
        return service.findAll();
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public AuthorDTO create(@RequestBody @Valid AuthorDTO dto) {
        return service.create(dto);
    }

    @DeleteMapping("/{id}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void delete(@PathVariable Long id) {
        service.delete(id);
    }
}
